Refactor the password generator and fix potential bugs.
- The original `password.py` script has been refactored into two separate modules: `generate_password.py` and `generate_multiple_passwords.py`, aligning the codebase with the `README.md`.
- Fixed a potential infinite loop in the password generation process by adding a retry limit.
- Fixed a potential `IndexError` by adding a check for the API response.
- Updated the deprecated "davinci" OpenAI model to "gpt-3.5-turbo-instruct".
